code/requirements.txt
asgiref==3.6.0
Django==4.1.4
mysql-connector-python==8.0.31
protobuf==3.20.1
sqlparse==0.4.3
tzdata==2022.7
